requirements.txt
Django>=1.4
MySQL-python>=1.2.3
pillow>=1.7.7
beautifulsoup4>=4.1.1
decorator>=3.3.3
django-allauth>=0.7.0
django-annoying>=0.7.6
django-celery>=3.0.4
django-gravatar2>=1.0.6
django-avatar>=1.0.5
django-notification>=0.2
django-socialauth>=0.1.2c
django-tastypie>=0.9.11
django-tinymce>=1.5.1b2
git+git://github.com/PaulUithol/backbone-tastypie.git#egg=backbone-tastypie
git+git://github.com/maraujop/django-crispy-forms.git#django-crispy-forms
django-grappelli
humanize